Single program for all three KP projects.
You need to change 1_param2.txt inside the Input folder to suit your project; set only ONE of Reward or Cost or Size to 1, and set the rest to 0.

Cost KP: You should see two additional scenes, one that shows a randomly generated number before the trial starts (currently from 1000 to 9999, but that's easily adjustable) and another scene to input the random number after the answer screen (submission can be done either through pressing enter on your keyboard or click "Submit").
Reward KP: You should see an addition scene at the start of each block; with the reward level for that block (you can change the reward level inside 1_param2.txt).
There are no major changes to the "Size KP", Terry's program works fine as is.

Also, items for Reward/Cost KP are presented in the form of a circle to make the game more suitable for eye tracking.

Note that we still need to adjust output files. I think you should speak with Nitin and Pablo to decide on the format of input files and which instances to use, before we work on the structure of output files.